# Conflicts:
#	.gitignore
#	CONTRIBUTING.md
#	LICENSE.txt
#	Library/.rubocop.yml
#	Library/Homebrew/README.md
#	Library/Homebrew/build.rb
#	Library/Homebrew/build_environment.rb
#	Library/Homebrew/build_options.rb
#	Library/Homebrew/caveats.rb
#	Library/Homebrew/checksum.rb
#	Library/Homebrew/cleaner.rb
#	Library/Homebrew/cmd/--cache.rb
#	Library/Homebrew/cmd/--cellar.rb
#	Library/Homebrew/cmd/--env.rb
#	Library/Homebrew/cmd/--prefix.rb
#	Library/Homebrew/cmd/--repository.rb
#	Library/Homebrew/cmd/cat.rb
#	Library/Homebrew/cmd/cleanup.rb
#	Library/Homebrew/cmd/command.rb
#	Library/Homebrew/cmd/commands.rb
#	Library/Homebrew/cmd/config.rb
#	Library/Homebrew/cmd/deps.rb
#	Library/Homebrew/cmd/desc.rb
#	Library/Homebrew/cmd/diy.rb
#	Library/Homebrew/cmd/doctor.rb
#	Library/Homebrew/cmd/fetch.rb
#	Library/Homebrew/cmd/gist-logs.rb
#	Library/Homebrew/cmd/help.rb
#	Library/Homebrew/cmd/home.rb
#	Library/Homebrew/cmd/info.rb
#	Library/Homebrew/cmd/install.rb
#	Library/Homebrew/cmd/leaves.rb
#	Library/Homebrew/cmd/link.rb
#	Library/Homebrew/cmd/list.rb
#	Library/Homebrew/cmd/log.rb
#	Library/Homebrew/cmd/migrate.rb
#	Library/Homebrew/cmd/missing.rb
#	Library/Homebrew/cmd/options.rb
#	Library/Homebrew/cmd/outdated.rb
#	Library/Homebrew/cmd/pin.rb
#	Library/Homebrew/cmd/postinstall.rb
#	Library/Homebrew/cmd/prune.rb
#	Library/Homebrew/cmd/readall.rb
#	Library/Homebrew/cmd/reinstall.rb
#	Library/Homebrew/cmd/search.rb
#	Library/Homebrew/cmd/sh.rb
#	Library/Homebrew/cmd/style.rb
#	Library/Homebrew/cmd/switch.rb
#	Library/Homebrew/cmd/tap-info.rb
#	Library/Homebrew/cmd/tap-pin.rb
#	Library/Homebrew/cmd/tap-unpin.rb
#	Library/Homebrew/cmd/tap.rb
#	Library/Homebrew/cmd/uninstall.rb
#	Library/Homebrew/cmd/unlink.rb
#	Library/Homebrew/cmd/unpack.rb
#	Library/Homebrew/cmd/unpin.rb
#	Library/Homebrew/cmd/untap.rb
#	Library/Homebrew/cmd/upgrade.rb
#	Library/Homebrew/cmd/uses.rb
#	Library/Homebrew/compat.rb
#	Library/Homebrew/compat/formula.rb
#	Library/Homebrew/compat/requirements.rb
#	Library/Homebrew/compilers.rb
#	Library/Homebrew/config.rb
#	Library/Homebrew/cxxstdlib.rb
#	Library/Homebrew/debrew.rb
#	Library/Homebrew/debrew/irb.rb
#	Library/Homebrew/dependable.rb
#	Library/Homebrew/dependencies.rb
#	Library/Homebrew/dependency.rb
#	Library/Homebrew/dependency_collector.rb
#	Library/Homebrew/descriptions.rb
#	Library/Homebrew/dev-cmd/update-test.rb
#	Library/Homebrew/download_strategy.rb
#	Library/Homebrew/exceptions.rb
#	Library/Homebrew/extend/ARGV.rb
#	Library/Homebrew/extend/ENV.rb
#	Library/Homebrew/extend/ENV/shared.rb
#	Library/Homebrew/extend/ENV/std.rb
#	Library/Homebrew/extend/ENV/super.rb
#	Library/Homebrew/extend/fileutils.rb
#	Library/Homebrew/extend/module.rb
#	Library/Homebrew/extend/pathname.rb
#	Library/Homebrew/extend/string.rb
#	Library/Homebrew/formula.rb
#	Library/Homebrew/formula_assertions.rb
#	Library/Homebrew/formula_cellar_checks.rb
#	Library/Homebrew/formula_installer.rb
#	Library/Homebrew/formula_pin.rb
#	Library/Homebrew/formula_support.rb
#	Library/Homebrew/formula_versions.rb
#	Library/Homebrew/formulary.rb
#	Library/Homebrew/global.rb
#	Library/Homebrew/hardware.rb
#	Library/Homebrew/install_renamed.rb
#	Library/Homebrew/keg.rb
#	Library/Homebrew/keg_relocate.rb
#	Library/Homebrew/language/go.rb
#	Library/Homebrew/language/haskell.rb
#	Library/Homebrew/language/java.rb
#	Library/Homebrew/language/python.rb
#	Library/Homebrew/metafiles.rb
#	Library/Homebrew/migrator.rb
#	Library/Homebrew/official_taps.rb
#	Library/Homebrew/options.rb
#	Library/Homebrew/os.rb
#	Library/Homebrew/os/mac.rb
#	Library/Homebrew/os/mac/version.rb
#	Library/Homebrew/os/mac/xcode.rb
#	Library/Homebrew/os/mac/xquartz.rb
#	Library/Homebrew/patch.rb
#	Library/Homebrew/pkg_version.rb
#	Library/Homebrew/postinstall.rb
#	Library/Homebrew/requirement.rb
#	Library/Homebrew/requirements.rb
#	Library/Homebrew/requirements/java_requirement.rb
#	Library/Homebrew/requirements/maximum_macos_requirement.rb
#	Library/Homebrew/requirements/osxfuse_requirement.rb
#	Library/Homebrew/requirements/tuntap_requirement.rb
#	Library/Homebrew/requirements/unsigned_kext_requirement.rb
#	Library/Homebrew/requirements/x11_requirement.rb
#	Library/Homebrew/resource.rb
#	Library/Homebrew/sandbox.rb
#	Library/Homebrew/software_spec.rb
#	Library/Homebrew/tab.rb
#	Library/Homebrew/tap.rb
#	Library/Homebrew/tap_constants.rb
#	Library/Homebrew/test.rb
#	Library/Homebrew/test/Gemfile
#	Library/Homebrew/test/Gemfile.lock
#	Library/Homebrew/utils.rb
#	Library/Homebrew/utils/fork.rb
#	Library/Homebrew/utils/git.rb
#	Library/Homebrew/utils/inreplace.rb
#	Library/Homebrew/utils/popen.rb
#	Library/Homebrew/version.rb
#	README.md
#	bin/brew

+++ Library/Homebrew/migrator.rb
@@ -1,13 +1,11 @@
-require "formula"
-require "formula_lock"
+require "lock_file"
 require "keg"
 require "tab"
-require "tap_migrations"
 
 class Migrator
   class MigrationNeededError < RuntimeError
     def initialize(formula)
-      super <<-EOS.undent
+      super <<~EOS
         #{formula.oldname} was renamed to #{formula.name} and needs to be migrated.
         Please run `brew migrate #{formula.oldname}`
       EOS
@@ -28,18 +26,15 @@ class Migrator
 
   class MigratorDifferentTapsError < RuntimeError
     def initialize(formula, tap)
-      msg = if tap == "Homebrew/homebrew"
+      msg = if tap.core_tap?
         "Please try to use #{formula.oldname} to refer the formula.\n"
       elsif tap
-        user, repo = tap.split("/")
-        repo.sub!("homebrew-", "")
-        "Please try to use fully-qualified #{user}/#{repo}/#{formula.oldname} to refer the formula.\n"
+        "Please try to use fully-qualified #{tap}/#{formula.oldname} to refer the formula.\n"
       end
 
-      super <<-EOS.undent
-      #{formula.name} from #{formula.tap} is given, but old name #{formula.oldname} was installed from #{tap ? tap : "path or url"}.
-
-      #{msg}To force migrate use `brew migrate --force #{formula.oldname}`.
+      super <<~EOS
+        #{formula.name} from #{formula.tap} is given, but old name #{formula.oldname} was installed from #{tap ? tap : "path or url"}.
+         #{msg}To force migrate use `brew migrate --force #{formula.oldname}`.
       EOS
     end
   end
@@ -80,38 +75,61 @@ class Migrator
   # path to newname cellar according to new name
   attr_reader :new_cellar
 
+  # true if new cellar existed at initialization time
+  attr_reader :new_cellar_existed
+
   # path to newname pin
   attr_reader :new_pin_record
 
   # path to newname keg that will be linked if old_linked_keg isn't nil
   attr_reader :new_linked_keg_record
 
-  def initialize(formula)
+  def self.needs_migration?(formula)
+    oldname = formula.oldname
+    return false unless oldname
+    oldname_rack = HOMEBREW_CELLAR/oldname
+    return false if oldname_rack.symlink?
+    return false unless oldname_rack.directory?
+    true
+  end
+
+  def self.migrate_if_needed(formula)
+    return unless Migrator.needs_migration?(formula)
+    begin
+      migrator = Migrator.new(formula)
+      migrator.migrate
+    rescue => e
+      onoe e
+    end
+  end
+
+  def initialize(formula, force: ARGV.force?)
     @oldname = formula.oldname
     @newname = formula.name
-    raise MigratorNoOldnameError.new(formula) unless oldname
+    raise MigratorNoOldnameError, formula unless oldname
 
     @formula = formula
     @old_cellar = HOMEBREW_CELLAR/formula.oldname
-    raise MigratorNoOldpathError.new(formula) unless old_cellar.exist?
+    raise MigratorNoOldpathError, formula unless old_cellar.exist?
 
     @old_tabs = old_cellar.subdirs.map { |d| Tab.for_keg(Keg.new(d)) }
     @old_tap = old_tabs.first.tap
 
-    if !ARGV.force? && !from_same_taps?
+    if !force && !from_same_tap_user?
       raise MigratorDifferentTapsError.new(formula, old_tap)
     end
 
     @new_cellar = HOMEBREW_CELLAR/formula.name
+    @new_cellar_existed = @new_cellar.exist?
 
-    if @old_linked_keg = get_linked_old_linked_keg
+    if @old_linked_keg = linked_old_linked_keg
       @old_linked_keg_record = old_linked_keg.linked_keg_record if old_linked_keg.linked?
       @old_opt_record = old_linked_keg.opt_record if old_linked_keg.optlinked?
       @new_linked_keg_record = HOMEBREW_CELLAR/"#{newname}/#{File.basename(old_linked_keg)}"
     end
 
-    @old_pin_record = HOMEBREW_LIBRARY/"PinnedKegs"/oldname
-    @new_pin_record = HOMEBREW_LIBRARY/"PinnedKegs"/newname
+    @old_pin_record = HOMEBREW_PINNED_KEGS/oldname
+    @new_pin_record = HOMEBREW_PINNED_KEGS/newname
     @pinned = old_pin_record.symlink?
     @old_pin_link_record = old_pin_record.readlink if @pinned
   end
@@ -119,31 +137,43 @@ class Migrator
   # Fix INSTALL_RECEIPTS for tap-migrated formula.
   def fix_tabs
     old_tabs.each do |tab|
-      tab.source["tap"] = formula.tap
+      tab.tap = formula.tap
       tab.write
     end
   end
 
-  def from_same_taps?
-    if formula.tap == old_tap
+  def from_same_tap_user?
+    formula_tap_user = formula.tap.user if formula.tap
+    old_tap_user = nil
+
+    new_tap = if old_tap
+      old_tap_user, = old_tap.user
+      if migrate_tap = old_tap.tap_migrations[formula.oldname]
+        new_tap_user, new_tap_repo = migrate_tap.split("/")
+        "#{new_tap_user}/#{new_tap_repo}"
+      end
+    end
+
+    if formula_tap_user == old_tap_user
       true
     # Homebrew didn't use to update tabs while performing tap-migrations,
-    # so there can be INSTALL_RECEIPT's containing wrong information about
-    # tap (tap is Homebrew/homebrew if installed formula migrates to a tap), so
-    # we check if there is an entry about oldname migrated to tap and if
+    # so there can be INSTALL_RECEIPT's containing wrong information about tap,
+    # so we check if there is an entry about oldname migrated to tap and if
     # newname's tap is the same as tap to which oldname migrated, then we
     # can perform migrations and the taps for oldname and newname are the same.
-    elsif TAP_MIGRATIONS && (rec = TAP_MIGRATIONS[formula.oldname]) \
-        && rec == formula.tap.sub("homebrew-", "") && old_tap == "Homebrew/homebrew"
+    elsif formula.tap && old_tap && formula.tap == new_tap
       fix_tabs
       true
-    elsif formula.tap
+    else
       false
     end
   end
 
-  def get_linked_old_linked_keg
-    kegs = old_cellar.subdirs.map { |d| Keg.new(d) }
+  def linked_old_linked_keg
+    keg_dirs = []
+    keg_dirs += new_cellar.subdirs if new_cellar.exist?
+    keg_dirs += old_cellar.subdirs
+    kegs = keg_dirs.map { |d| Keg.new(d) }
     kegs.detect(&:linked?) || kegs.detect(&:optlinked?)
   end
 
@@ -152,69 +182,101 @@ class Migrator
   end
 
   def migrate
-    if new_cellar.exist?
-      onoe "#{new_cellar} already exists; remove it manually and run brew migrate #{oldname}."
-      return
-    end
-
-    begin
-      oh1 "Migrating #{Tty.green}#{oldname}#{Tty.white} to #{Tty.green}#{newname}#{Tty.reset}"
-      lock
-      unlink_oldname
-      move_to_new_directory
-      repin
-      link_newname unless old_linked_keg.nil?
-      link_oldname_opt
-      link_oldname_cellar
-      update_tabs
-    rescue Interrupt
-      ignore_interrupts { backup_oldname }
-    rescue Exception => e
-      onoe "Error occured while migrating."
-      puts e
-      puts e.backtrace if ARGV.debug?
-      puts "Backuping..."
-      ignore_interrupts { backup_oldname }
-    ensure
-      unlock
-    end
+    oh1 "Processing #{Formatter.identifier(oldname)} formula rename to #{Formatter.identifier(newname)}"
+    lock
+    unlink_oldname
+    unlink_newname if new_cellar.exist?
+    repin
+    move_to_new_directory
+    link_oldname_cellar
+    link_oldname_opt
+    link_newname unless old_linked_keg.nil?
+    update_tabs
+    return unless formula.outdated?
+    opoo <<~EOS
+      #{Formatter.identifier(newname)} is outdated!
+      To avoid broken installations, as soon as possible please run:
+        brew upgrade
+      Or, if you're OK with a less reliable fix:
+        brew upgrade #{newname}
+    EOS
+  rescue Interrupt
+    ignore_interrupts { backup_oldname }
+  rescue Exception => e # rubocop:disable Lint/RescueException
+    onoe "Error occurred while migrating."
+    puts e
+    puts e.backtrace if ARGV.debug?
+    puts "Backing up..."
+    ignore_interrupts { backup_oldname }
+  ensure
+    unlock
   end
 
   # move everything from Cellar/oldname to Cellar/newname
   def move_to_new_directory
-    puts "Moving to: #{new_cellar}"
-    FileUtils.mv(old_cellar, new_cellar)
-  end
+    return unless old_cellar.exist?
 
-  def repin
-    if pinned?
-      # old_pin_record is a relative symlink and when we try to to read it
-      # from <dir> we actually try to find file
-      # <dir>/../<...>/../Cellar/name/version.
-      # To repin formula we need to update the link thus that it points to
-      # the right directory.
-      # NOTE: old_pin_record.realpath.sub(oldname, newname) is unacceptable
-      # here, because it resolves every symlink for old_pin_record and then
-      # substitutes oldname with newname. It breaks things like
-      # Pathname#make_relative_symlink, where Pathname#relative_path_from
-      # is used to find relative path from source to destination parent and
-      # it assumes no symlinks.
-      src_oldname = old_pin_record.dirname.join(old_pin_link_record).expand_path
-      new_pin_record.make_relative_symlink(src_oldname.sub(oldname, newname))
-      old_pin_record.delete
+    if new_cellar.exist?
+      conflicted = false
+      old_cellar.each_child do |c|
+        next unless (new_cellar/c.basename).exist?
+        begin
+          FileUtils.rm_rf c
+        rescue Errno::EACCES
+          conflicted = true
+          onoe "#{new_cellar/c.basename} already exists."
+        end
+      end
+
+      if conflicted
+        odie "Remove #{new_cellar} manually and run brew migrate #{oldname}."
+      end
+    end
+
+    oh1 "Moving #{Formatter.identifier(oldname)} versions to #{new_cellar}"
+    if new_cellar.exist?
+      FileUtils.mv(old_cellar.children, new_cellar)
+    else
+      FileUtils.mv(old_cellar, new_cellar)
     end
   end
 
+  def repin
+    return unless pinned?
+    # old_pin_record is a relative symlink and when we try to to read it
+    # from <dir> we actually try to find file
+    # <dir>/../<...>/../Cellar/name/version.
+    # To repin formula we need to update the link thus that it points to
+    # the right directory.
+    # NOTE: old_pin_record.realpath.sub(oldname, newname) is unacceptable
+    # here, because it resolves every symlink for old_pin_record and then
+    # substitutes oldname with newname. It breaks things like
+    # Pathname#make_relative_symlink, where Pathname#relative_path_from
+    # is used to find relative path from source to destination parent and
+    # it assumes no symlinks.
+    src_oldname = (old_pin_record.dirname/old_pin_link_record).expand_path
+    new_pin_record.make_relative_symlink(src_oldname.sub(oldname, newname))
+    old_pin_record.delete
+  end
+
   def unlink_oldname
-    oh1 "Unlinking #{Tty.green}#{oldname}#{Tty.reset}"
+    oh1 "Unlinking #{Formatter.identifier(oldname)}"
     old_cellar.subdirs.each do |d|
       keg = Keg.new(d)
       keg.unlink
     end
   end
 
+  def unlink_newname
+    oh1 "Temporarily unlinking #{Formatter.identifier(newname)}"
+    new_cellar.subdirs.each do |d|
+      keg = Keg.new(d)
+      keg.unlink
+    end
+  end
+
   def link_newname
-    oh1 "Linking #{Tty.green}#{newname}#{Tty.reset}"
+    oh1 "Relinking #{Formatter.identifier(newname)}"
     new_keg = Keg.new(new_linked_keg_record)
 
     # If old_keg wasn't linked then we just optlink a keg.
@@ -233,13 +295,14 @@ class Migrator
     new_keg.remove_linked_keg_record if new_keg.linked?
 
     begin
-      new_keg.link
+      mode = OpenStruct.new(overwrite: true)
+      new_keg.link(mode)
     rescue Keg::ConflictError => e
       onoe "Error while executing `brew link` step on #{newname}"
       puts e
       puts
       puts "Possible conflicting files are:"
-      mode = OpenStruct.new(:dry_run => true, :overwrite => true)
+      mode = OpenStruct.new(dry_run: true, overwrite: true)
       new_keg.link(mode)
       raise
     rescue Keg::LinkError => e
@@ -248,7 +311,7 @@ class Migrator
       puts
       puts "You can try again using:"
       puts "  brew link #{formula.name}"
-    rescue Exception => e
+    rescue Exception => e # rubocop:disable Lint/RescueException
       onoe "An unexpected error occurred during linking"
       puts e
       puts e.backtrace if ARGV.debug?
@@ -259,10 +322,9 @@ class Migrator
 
   # Link keg to opt if it was linked before migrating.
   def link_oldname_opt
-    if old_opt_record
-      old_opt_record.delete if old_opt_record.symlink?
-      old_opt_record.make_relative_symlink(new_linked_keg_record)
-    end
+    return unless old_opt_record
+    old_opt_record.delete if old_opt_record.symlink?
+    old_opt_record.make_relative_symlink(new_linked_keg_record)
   end
 
   # After migtaion every INSTALL_RECEIPT.json has wrong path to the formula
@@ -300,7 +362,7 @@ class Migrator
     end
   end
 
-  # Backup everything if errors occured while migrating.
+  # Backup everything if errors occurred while migrating.
   def backup_oldname
     unlink_oldname_opt
     unlink_oldname_cellar
@@ -308,7 +370,7 @@ class Migrator
     backup_old_tabs
 
     if pinned? && !old_pin_record.symlink?
-      src_oldname = old_pin_record.dirname.join(old_pin_link_record).expand_path
+      src_oldname = (old_pin_record.dirname/old_pin_link_record).expand_path
       old_pin_record.make_relative_symlink(src_oldname)
       new_pin_record.delete
     end
@@ -317,34 +379,31 @@ class Migrator
       new_cellar.subdirs.each do |d|
         newname_keg = Keg.new(d)
         newname_keg.unlink
-        newname_keg.uninstall
+        newname_keg.uninstall if new_cellar_existed
       end
     end
 
-    unless old_linked_keg.nil?
-      # The keg used to be linked and when we backup everything we restore
-      # Cellar/oldname, the target also gets restored, so we are able to
-      # create a keg using its old path
-      if old_linked_keg_record
-        begin
-          old_linked_keg.link
-        rescue Keg::LinkError
-          old_linked_keg.unlink
-          raise
-        rescue Keg::AlreadyLinkedError
-          old_linked_keg.unlink
-          retry
-        end
-      else
-        old_linked_keg.optlink
+    return if old_linked_keg.nil?
+    # The keg used to be linked and when we backup everything we restore
+    # Cellar/oldname, the target also gets restored, so we are able to
+    # create a keg using its old path
+    if old_linked_keg_record
+      begin
+        old_linked_keg.link
+      rescue Keg::LinkError
+        old_linked_keg.unlink
+        raise
+      rescue Keg::AlreadyLinkedError
+        old_linked_keg.unlink
+        retry
       end
+    else
+      old_linked_keg.optlink
     end
   end
 
   def backup_oldname_cellar
-    unless old_cellar.exist?
-      FileUtils.mv(new_cellar, old_cellar)
-    end
+    FileUtils.mv(new_cellar, old_cellar) unless old_cellar.exist?
   end
 
   def backup_old_tabs
